my folder containing all my codes for the web extension
This web extension helps to get the HTTP's within any web page that is open. you just have to load the web page and then click on the extension to generate for you all the HTTP's within that very web page. you can try it with " muele.mak.ac.ug"